+++ test/heterogeneous-test.js
@@ -6,18 +6,6 @@ var tape = require("tape"),
  ************ Components *************
  *************************************/
 
-// Basic components and nesting.
-var heading = d3.component("h1")
-      .render(function (selection, d){ selection.text(d); }),
-    paragraph = d3.component("p")
-      .render(function (selection, d){ selection.text(d); }),
-    post = d3.component("div")
-      .render(function (selection, d){
-        selection
-          .call(heading, d.title)
-          .call(paragraph, d.content);
-      });
-
 // Heterogeneous component trees.
 var apple = d3.component("span", "apple")
       .render(function (selection, d){ selection.text(d); }),
@@ -36,107 +24,9 @@ var apple = d3.component("span", "apple")
           .call(orange, d.oranges || []);
       });
 
-// Local state.
-var spinnerCreated = false,
-    spinnerDestroyed = false,
-    spinner = d3.component("div")
-      .create(function (setState){
-        setState({
-          timer: "running",
-          cleanup: function (){
-            setState({
-              timer: "stopped"
-            });
-          }
-        });
-      })
-      .render(function (selection, props, state){
-        selection.text("Timer is " + state.timer);
-      }
-      .destroy(function(state){
-        state.cleanup();
-      });
-
 /*************************************
  ************** Tests ****************
  *************************************/
-tape("A component should render a single instance.", function(test) {
-  var div = createDiv();
-  div.call(paragraph, "foo");
-  test.equal(div.html(), "<p>foo</p>");
-  test.end();
-});
-
-
-tape("A component should render multiple instances.", function(test) {
-  var div = createDiv();
-
-  // Enter
-  div.call(paragraph, [ "foo", "bar" ]);
-  test.equal(div.html(), "<p>foo</p><p>bar</p>");
-
-  // Update + Enter
-  div.call(paragraph, [ "fooz", "barz", "baz" ]);
-  test.equal(div.html(), "<p>fooz</p><p>barz</p><p>baz</p>");
-
-  // Exit
-  div.call(paragraph, []);
-  test.equal(div.html(), "");
-
-  test.end();
-});
-
-
-tape("Nested components.", function(test) {
-  var div = createDiv();
-
-  div.call(post, {
-    title: "Title",
-    content: "Content here."
-  });
-
-  test.equal(div.html(), [
-    "<div>",
-      "<h1>Title</h1>",
-      "<p>Content here.</p>",
-    "</div>"
-  ].join(""));
-
-  test.end();
-});
-
-
-tape("Nested components multiple instances.", function(test) {
-  var div = createDiv();
-
-  // Enter
-  div.call(post, [
-    { title: "A", content: "a" },
-    { title: "B", content: "b" },
-  ]);
-  test.equal(div.html(), [
-    "<div><h1>A</h1><p>a</p></div>",
-    "<div><h1>B</h1><p>b</p></div>"
-  ].join(""));
-
-  // Enter + Update
-  div.call(post, [
-    { title: "D", content: "d" },
-    { title: "E", content: "e" },
-    { title: "F", content: "f" },
-  ]);
-  test.equal(div.html(), [
-    "<div><h1>D</h1><p>d</p></div>",
-    "<div><h1>E</h1><p>e</p></div>",
-    "<div><h1>F</h1><p>f</p></div>"
-  ].join(""));
-
-  // Exit
-  div.call(post, []);
-  test.equal(div.html(), "");
-
-  test.end();
-});
 
 tape("Conditional components with classes.", function(test) {
   var div = createDiv();
@@ -202,13 +92,6 @@ tape("Multiple nested component types.", function(test) {
       '<span class="orange">Mandarin</span>',
     "</div>"
   ].join(""));
-
-  test.end();
-});
-
-
-tape("Local state.", function(test) {
-  var div = createDiv();
 
   test.end();
 });
